+++ src/Miningcore/Blockchain/Kaspa/Custom/Spectre/SpectreJob.cs
@@ -8,7 +8,7 @@ using Miningcore.Native;
 using Miningcore.Stratum;
 using Miningcore.Util;
 using NBitcoin;
-using kaspad = Miningcore.Blockchain.Kaspa.Kaspad;
+using Kaspad = Miningcore.Blockchain.Kaspa.Kaspad;
 
 namespace Miningcore.Blockchain.Kaspa.Custom.Spectre;
 
@@ -109,7 +109,7 @@ public class SpectreJob : KaspaJob
         return result;
     }
 
-    public override void Init(kaspad.RpcBlock blockTemplate, string jobId, double shareMultiplier)
+    public override void Init(Kaspad.RpcBlock blockTemplate, string jobId, double shareMultiplier)
     {
         Contract.RequiresNonNull(blockTemplate);
         Contract.Requires<ArgumentException>(!string.IsNullOrEmpty(jobId));